requirements.txt
absl-py==2.2.2
accelerate==0.32.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anndata==0.10.9
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
array_api_compat==1.11.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
build==1.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
chromadb==1.5.2
click==8.1.8
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.4.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
diffusers==0.28.2
dill==0.4.0
distro==1.9.0
durationpy==0.10
evaluate==0.4.6
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.13.0
fastapi==0.128.8
ffmpy==1.0.0
filelock==3.13.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.2.10
fonttools==4.48.1
frozenlist==1.8.0
fsspec==2024.2.0
gast==0.6.0
get-annotations==0.1.2
git-filter-repo==2.47.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.2.4
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.2.0
httpcore==1.0.7
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
ibm-cos-sdk==2.16.0
ibm-cos-sdk-core==2.16.0
ibm-cos-sdk-s3transfer==2.16.0
idna==3.6
imageio==2.31.3
importlib-resources==6.1.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719361860083/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1701831663892/work
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257447442/work
jupyterlab_widgets==3.0.11
kagglehub==0.3.12
keras==3.10.0
kiwisolver==1.4.5
kubernetes==35.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.37
legacy-api-wrap==1.5
libclang==18.1.1
lightning-utilities==0.15.2
llama_stack==0.0.1a5
llvmlite==0.43.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.9.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mkl-service==2.5.2
mkl_fft @ file:///croot/mkl_fft_1751317925336/work
mkl_random @ file:///home/task_176159294118956/conda-bld/mkl_random_1761592954803/work
ml_dtypes==0.5.1
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
mvlearn==0.5.0
mypy_extensions==1.1.0
namex==0.0.8
natsort==8.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
nltk==3.9.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.19.2
openai==1.109.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-python==4.11.0.86
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.15.0
orjson==3.11.5
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.1.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
patsy==1.0.2
peft==0.11.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
polars==1.36.1
polars-runtime-32==1.36.1
posthog==5.4.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
propcache==0.4.1
protobuf==5.29.4
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1719274564771/work
ptflops==0.7.4
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==21.0.0
pybase64==1.4.3
pydantic==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.27.2
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyMuPDF==1.26.5
pynndescent==0.6.0
pyparsing==3.1.1
pypdf==6.7.5
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1705605076900/work
ray==2.51.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge-score==0.1.2
rpds-py==0.27.1
ruff==0.15.1
sacremoses==0.1.1
safetensors==0.7.0
scanpy==1.10.3
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
sentence-transformers==2.7.0
session_info==1.0.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
skillsnetwork==0.21.10
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.49.3
statsmodels==0.14.6
stdlib-list==0.12.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.20.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tf_keras==2.20.1
threadpoolctl==3.2.0
tiktoken==0.12.0
timm==0.3.2
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.12.0
torch==2.8.0
torch-cka==0.21
torchaudio==2.8.0
torcheval==0.0.7
torchmetrics==1.8.2
torchvision==0.23.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1717722887379/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.57.6
triton==3.4.0
trl==0.4.7
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
ultralytics==8.4.6
ultralytics-thop==2.0.18
umap-learn==0.5.11
urllib3==2.6.2
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websocket-client==1.9.0
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.11
wrapt==1.17.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.17.0
zstandard==0.25.0
